[BOND-8207] lib updates (#40)
* lib updates

* docker pulsar update

* docker elastic update

* Adding Pulsar Generated Admin

* disable transactions

* cleanup

* namespace change

* remove SharpPulsar.Admin

* fix data format

* disable topic creation

* sharp admin

* disable transactions

* ignore delete error if 200

* ignore 404

* build fix

* log docker

* add admin panel

* downgrade docker compose

* one at a time

* fixing elastic

* Revert "one at a time"

This reverts commit fd202068bb0664b0852a1b2a69da4cb3df0b6760.

* Revert "downgrade docker compose"

This reverts commit 582796cfa16f015b026bf5919943610a436f2351.

* Revert "log docker"

This reverts commit a4fa8adc84a101ed48714a1d544e4800b2a1ea08.

* delete topic patch

* ignore 409 for tenant and namespace

* formatter test fix

* formatter test category unit

* test fix

* add Pulsar Exception

* add logs

* cleanup

* aggregator build publisher

* Adding Defensive check and log

* trigger aggregator

* cleanup

* check ex

+++ src/EventHorizon.EventSourcing/Aggregates/Aggregator.cs
@@ -10,6 +10,7 @@ using EventHorizon.Abstractions.Interfaces;
 using EventHorizon.Abstractions.Interfaces.Internal;
 using EventHorizon.Abstractions.Models;
 using EventHorizon.Abstractions.Models.TopicMessages;
+using EventHorizon.Abstractions.Reflection;
 using EventHorizon.Abstractions.Serialization.Compression.Extensions;
 using EventHorizon.EventStore;
 using EventHorizon.EventStore.Interfaces;
@@ -30,8 +31,9 @@ public class Aggregator<TParent, TState>
     private readonly StreamingClient _streamingClient;
     private readonly AggregatorConfig<TState> _config;
     private readonly Dictionary<string, object> _publisherDict = new();
-    private readonly string _eventTopic;
     private readonly Store<TParent, TState> _store;
+    private readonly StateDetail _stateDetail;
+    private readonly Publisher<Event> _publisher;
 
     public Aggregator(
         ICrudStore<TParent> crudStore,
@@ -43,7 +45,8 @@ public class Aggregator<TParent, TState>
         _store = new StoreBuilder<TParent, TState>(crudStore).AddCompression(config.StateCompression).Build();
         _streamingClient = streamingClient;
         _config = config;
-        _eventTopic = formatter.GetTopic<Event>(_stateType);
+        _publisher = GetPublisher<Event>(formatter.GetTopic<Event>(_stateType));
+        _stateDetail = ReflectionFactory.GetStateDetail(_stateType);
         _logger = logger;
     }
 
@@ -119,8 +122,7 @@ public class Aggregator<TParent, TState>
 
         try
         {
-            var publisher = GetPublisher<Event>(_eventTopic);
-            await publisher.PublishAsync(events).ConfigureAwait(false);
+            await _publisher.PublishAsync(events).ConfigureAwait(false);
         }
         catch (Exception ex)
         {
@@ -181,7 +183,6 @@ public class Aggregator<TParent, TState>
         try
         {
             // Load Snapshots
-            var sw = Stopwatch.StartNew();
             streamIds = streamIds.Distinct().ToArray();
             var snapshots = await _store.GetAllAsync(streamIds, ct).ConfigureAwait(false);
             var parentDict = snapshots.ToDictionary(x => x.Id);
@@ -245,8 +246,12 @@ public class Aggregator<TParent, TState>
     {
         await _store.DropDatabaseAsync(ct).ConfigureAwait(false);
         await _streamingClient.GetAdmin<Event>().DeleteTopicAsync(_stateType, ct: ct).ConfigureAwait(false);
-        await _streamingClient.GetAdmin<Command>().DeleteTopicAsync(_stateType, ct: ct).ConfigureAwait(false);
-        await _streamingClient.GetAdmin<Request>().DeleteTopicAsync(_stateType, ct: ct).ConfigureAwait(false);
+
+        if(_stateDetail.HandlerDict[typeof(Command)].Count != 0)
+            await _streamingClient.GetAdmin<Command>().DeleteTopicAsync(_stateType, ct: ct).ConfigureAwait(false);
+
+        if(_stateDetail.HandlerDict[typeof(Request)].Count != 0)
+            await _streamingClient.GetAdmin<Request>().DeleteTopicAsync(_stateType, ct: ct).ConfigureAwait(false);
     }
 
     #endregion
